+++ src/main/java/com/vaadin/flow/component/textfield/GeneratedVaadinPasswordField.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.textfield;
+package com.vaadin.flow.component.textfield;
 
 import javax.annotation.Generated;
 import com.vaadin.ui.Tag;
