+++ src/test/java/com/vaadin/flow/component/textfield/demo/PasswordFieldView.java
@@ -13,15 +13,15 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.textfield.demo;
+package com.vaadin.flow.component.textfield.demo;
 
+import com.vaadin.flow.component.html.Div;
+import com.vaadin.flow.component.html.NativeButton;
+import com.vaadin.flow.component.textfield.GeneratedVaadinPasswordField;
+import com.vaadin.flow.component.textfield.PasswordField;
 import com.vaadin.flow.demo.DemoView;
 import com.vaadin.router.Route;
 import com.vaadin.ui.common.HtmlImport;
-import com.vaadin.ui.html.Div;
-import com.vaadin.ui.html.NativeButton;
-import com.vaadin.ui.textfield.GeneratedVaadinPasswordField;
-import com.vaadin.ui.textfield.PasswordField;
 
 /**
  * View for {@link GeneratedVaadinPasswordField} demo.
