+++ src/test/java/com/vaadin/flow/component/textfield/demo/TextAreaView.java
@@ -13,14 +13,14 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.textfield.demo;
+package com.vaadin.flow.component.textfield.demo;
 
+import com.vaadin.flow.component.html.Div;
+import com.vaadin.flow.component.textfield.GeneratedVaadinTextArea;
+import com.vaadin.flow.component.textfield.TextArea;
 import com.vaadin.flow.demo.DemoView;
 import com.vaadin.router.Route;
 import com.vaadin.ui.common.HtmlImport;
-import com.vaadin.ui.html.Div;
-import com.vaadin.ui.textfield.GeneratedVaadinTextArea;
-import com.vaadin.ui.textfield.TextArea;
 
 /**
  * View for {@link GeneratedVaadinTextArea} demo.
