+++ src/test/java/com/vaadin/flow/component/textfield/demo/TextFieldView.java
@@ -13,18 +13,18 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.textfield.demo;
+package com.vaadin.flow.component.textfield.demo;
 
+import com.vaadin.flow.component.html.Div;
+import com.vaadin.flow.component.textfield.GeneratedVaadinTextField;
+import com.vaadin.flow.component.textfield.TextField;
 import com.vaadin.flow.demo.DemoView;
 import com.vaadin.router.Route;
 import com.vaadin.ui.common.HtmlImport;
-import com.vaadin.ui.html.Div;
-import com.vaadin.ui.textfield.GeneratedVaadinTextField;
-import com.vaadin.ui.textfield.TextField;
 
 /**
  * View for {@link GeneratedVaadinTextField} demo.
- * 
+ *
  * @author Vaadin Ltd
  */
 @Route("vaadin-text-field")
