+++ src/test/java/com/vaadin/flow/component/textfield/tests/PasswordFieldPageIT.java
@@ -13,23 +13,23 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.textfield.tests;
+package com.vaadin.flow.component.textfield.tests;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.openqa.selenium.WebElement;
 
+import com.vaadin.flow.component.textfield.PasswordField;
 import com.vaadin.flow.testutil.AbstractComponentIT;
 import com.vaadin.flow.testutil.TestPath;
 import com.vaadin.testbench.By;
-import com.vaadin.ui.textfield.TextField;
 
 /**
- * Integration tests for {@link TextField}.
+ * Integration tests for {@link PasswordField}.
  */
-@TestPath("text-field-test")
-public class TextFieldPageIT extends AbstractComponentIT {
+@TestPath("password-field-test")
+public class PasswordFieldPageIT extends AbstractComponentIT {
 
     @Before
     public void init() {
@@ -38,7 +38,8 @@ public class TextFieldPageIT extends AbstractComponentIT {
 
     @Test
     public void assertReadOnly() {
-        WebElement webComponent = findElement(By.tagName("vaadin-text-field"));
+        WebElement webComponent = findElement(
+                By.tagName("vaadin-password-field"));
 
         Assert.assertNull(webComponent.getAttribute("readonly"));
 
@@ -56,7 +57,8 @@ public class TextFieldPageIT extends AbstractComponentIT {
 
     @Test
     public void assertRequired() {
-        WebElement webComponent = findElement(By.tagName("vaadin-text-field"));
+        WebElement webComponent = findElement(
+                By.tagName("vaadin-password-field"));
 
         Assert.assertNull(webComponent.getAttribute("required"));
 
@@ -68,18 +70,5 @@ public class TextFieldPageIT extends AbstractComponentIT {
         button.click();
         waitUntil(driver -> "false"
                 .equals(getProperty(webComponent, "required")));
-    }
-
-    @Test
-    public void assertValueWithoutListener() throws InterruptedException {
-        WebElement field = findElement(By.id("value-change"));
-
-        WebElement input = getInShadowRoot(field, By.cssSelector("input"));
-        input.sendKeys("foo");
-
-        findElement(By.id("get-value")).click();
-
-        String value = findElement(By.className("text-field-value")).getText();
-        Assert.assertEquals("foo", value);
     }
 }
