+++ src/test/java/com/vaadin/flow/component/textfield/tests/PasswordFieldValidationPage.java
@@ -13,12 +13,12 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.textfield.tests;
+package com.vaadin.flow.component.textfield.tests;
 
+import com.vaadin.flow.component.textfield.PasswordField;
 import com.vaadin.flow.testutil.ValidationTestView;
 import com.vaadin.router.Route;
 import com.vaadin.ui.common.HasValidation;
-import com.vaadin.ui.textfield.PasswordField;
 
 /**
  * View for testing validation with {@link PasswordField}.
