+++ src/test/java/com/vaadin/flow/component/textfield/tests/TextAreaIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.textfield.tests;
+package com.vaadin.flow.component.textfield.tests;
 
 import java.util.stream.IntStream;
 
@@ -22,9 +22,9 @@ import org.junit.Test;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebElement;
 
+import com.vaadin.flow.component.textfield.TextArea;
 import com.vaadin.flow.demo.ComponentDemoTest;
 import com.vaadin.testbench.By;
-import com.vaadin.ui.textfield.TextArea;
 
 /**
  * Integration tests for the {@link TextArea}.
